requirements_fixed.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-yasg==1.21.7

# Environment
python-decouple==3.8

# Caching and Background Tasks
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# File handling
Pillow==10.0.1
whitenoise==6.6.0

# Production server
gunicorn==21.2.0
